remodelling structure
