+++ lib/passbox.rb
@@ -1,8 +1,11 @@
 require 'passbox/aes'
 require 'passbox/auth'
 require 'passbox/init'
-require 'passbox/crud'
-require 'passbox/create'
+require 'passbox/list'
+require 'passbox/crud/create'
+require 'passbox/crud/read'
+require 'passbox/crud/update'
+require 'passbox/crud/delete'
 require 'passbox/options'
 require 'passbox/colourize'
 include Passbox
