+++ lib/passbox/colourize.rb
@@ -3,7 +3,7 @@ class String
     def black;          "\e[30m#{self}\e[0m" end
     def red;            "\e[31m#{self}\e[0m" end
     def green;          "\e[32m#{self}\e[0m" end
-    def brown;          "\e[33m#{self}\e[0m" end
+    def yellow;          "\e[33m#{self}\e[0m" end
     def blue;           "\e[34m#{self}\e[0m" end
     def magenta;        "\e[35m#{self}\e[0m" end
     def cyan;           "\e[36m#{self}\e[0m" end
