+++ lib/passbox/crud/read.rb
@@ -1,0 +1,14 @@
+module Passbox
+
+    def read_pass
+        check_passbox
+        acc=verify_account
+        key = passbox_auth
+        if key
+            data = JSON.parse(decrypt("#{$pbdir}/#{acc}.pb", key))
+            print "username : #{data['username']}\n"
+            print "password : #{data['password']}\n"
+        end
+    end
+
+end
